fix: usar Decimal/Numeric en vez de float para montos de dinero
Los montos (total, cobrado, saldo, costo, cargo, abono, etc.) ahora se
manejan como Decimal en Python y se guardan en columnas Numeric(10, 2)
en SQLite, en vez de float/Float. float no puede representar
exactamente la mayoria de los montos decimales, y sumar o restar
dinero repetidamente (como pasa en cada pago, anulacion o devolucion)
puede acumular errores de redondeo con el tiempo. Decimal evita ese
problema de raiz en vez de depender de round(x, 2) despues de cada
operacion.

Cambios:
- Nuevo helper redondear_monto() en services/comun.py: convierte
  cualquier valor a Decimal con 2 decimales exactos.
- models/finanzas.py y models/cita.py: columnas de dinero de Float a
  Numeric(10, 2).
- schemas/finanzas.py y schemas/cita.py: campos de dinero (monto,
  costo, montoPagado) de float a Decimal en los payloads Pydantic.
- services/finanzas.py y services/citas.py: toda la aritmetica de
  dinero reescrita con Decimal en vez de round(float(x), 2).
- Nuevo tests/test_precision_monetaria.py: prueba de regresion que
  llama la funcion real registrar_pago() 500 veces seguidas y confirma
  que el saldo final es exacto.

No requiere migracion de base de datos: SQLite no impone tipos de
columna de forma estricta, asi que las bases de datos existentes (con
la columna declarada como FLOAT) siguen funcionando sin cambios; los
valores viejos se siguen leyendo bien y todo lo que se escriba de aca
en adelante queda exacto. Verificado con una base de datos de prueba
creada con el esquema anterior.

70 tests existentes + 2 nuevos pasan. Lint y format (ruff) limpios.

+++ backend/app/models/finanzas.py
@@ -1,4 +1,4 @@
-from sqlalchemy import JSON, Column, Float, Integer, String, Text
+from sqlalchemy import JSON, Column, Integer, Numeric, String, Text
 
 from ..database import Base
 
@@ -11,17 +11,17 @@ class PagoDB(Base):
     citaId = Column(Integer, index=True)
     concepto = Column(String(200))
     fecha = Column(String(50))
-    total = Column(Float)
-    cobrado = Column(Float)
-    saldo = Column(Float)
+    total = Column(Numeric(10, 2))
+    cobrado = Column(Numeric(10, 2))
+    saldo = Column(Numeric(10, 2))
     metodo = Column(String(50))
     tipoPago = Column(String(50))
     cuotas = Column(JSON)
     creadoEn = Column(String(50))
     fechaUltPago = Column(String(50))
     nota = Column(Text)
-    devuelto = Column(Float)
-    creditoFavor = Column(Float)
+    devuelto = Column(Numeric(10, 2))
+    creditoFavor = Column(Numeric(10, 2))
 
 
 class PlanDB(Base):
@@ -32,7 +32,7 @@ class PlanDB(Base):
     nombre = Column(String(150))
     tipo = Column(String(100))
     duracion = Column(String(50))
-    costo = Column(Float)
+    costo = Column(Numeric(10, 2))
     nSesiones = Column(Integer)
     descripcion = Column(Text)
     estado = Column(String(50))
@@ -47,14 +47,14 @@ class PlanPagoDB(Base):
     pagoId = Column(Integer, index=True)
     citaId = Column(Integer, index=True)
     concepto = Column(String(200))
-    totalAcordado = Column(Float)
-    anticipo = Column(Float)
+    totalAcordado = Column(Numeric(10, 2))
+    anticipo = Column(Numeric(10, 2))
     metodoPreferido = Column(String(50))
     estado = Column(String(50))
     cuotas = Column(JSON)
-    totalCuotas = Column(Float)
-    cobrado = Column(Float)
-    saldo = Column(Float)
+    totalCuotas = Column(Numeric(10, 2))
+    cobrado = Column(Numeric(10, 2))
+    saldo = Column(Numeric(10, 2))
     fechaCreacion = Column(String(50))
     creadoEn = Column(String(50))
 
@@ -68,8 +68,8 @@ class MovimientoCuentaDB(Base):
     pagoId = Column(Integer, nullable=True, index=True)
     tipo = Column(String(50), index=True)
     descripcion = Column(String(250))
-    cargo = Column(Float, default=0)
-    abono = Column(Float, default=0)
+    cargo = Column(Numeric(10, 2), default=0)
+    abono = Column(Numeric(10, 2), default=0)
     fecha = Column(String(50), index=True)
     metodo = Column(String(100))
     referencia = Column(String(150))
